Add test that opens all ArticlePanel views.
